.divert is a horrible idea, but here for posterity.
a => c with a diversion within to b breaks assumptions about when a
callback of .to('b', cb) should be fired (when the state changes at all,
or when c is finally reached?). It should really be a .on('c', cb)
instead.

a => c with b diversion was intended as a shortcut. For example,

ready => playing (shortcut)
	if (!loaded) divert('loading');
ready => loading
loading => playing

But instead it's probably best to be explicit when the change will be
requested?

if (current() === 'ready') {
	to('loading', function(err) {
		if (err) throw err;
		to('playing');
	})
}